Moved creation out of callback
