+++ lib/photoUtils.js
@@ -1,10 +1,9 @@
-const { BlobServiceClient, ContainerSASPermissions } = require('@azure/storage-blob')
-import { createFinancialSummarie } from '../airtable/request';
-
 const atob = require('atob');
-
 const azure = require('azure-storage');
 const blobService = azure.createBlobService();
+
+STORAGE_ACCOUNT_NAME = 'meepanyar';
+STORAGE_CONTAINER_NAME = 'bankslips';
 
 export function dataURIToBuffer(dataURI) {
     // convert base64 to raw binary data held in a string
@@ -35,34 +34,22 @@ export async function uploadBlobAndCreateSummary(blobName, blobString) {
 
     try {
         // Regex was taken from https://stackoverflow.com/a/39590768
-        let matches = blobString.match(/^data:([A-Za-z-+\/]+);base64,(.+)$/);
+        const matches = blobString.match(/^data:([A-Za-z-+\/]+);base64,(.+)$/);
         const type = matches[1];
         const buffer = Buffer.from(matches[2], 'base64');
 
-        let bankSlipURL = ''
+        const bankSlipURL = `https://${STORAGE_ACCOUNT_NAME}.blob.core.windows.net/${STORAGE_CONTAINER_NAME}/${blobName}`;
         await blobService.createBlockBlobFromText('bankslips', blobName, buffer, {
             contentType: type,
-        }, async function (error, result, response) {
-
-            if (!error) {
-                console.log("Blob uploaded");
-                bankSlipURL = `https://meepanyar.blob.core.windows.net/bankslips/${blobName}`
-
-                let financialSummaryPayload = {
-                }
-
-                if (bankSlipURL) {
-                    financialSummaryPayload.bankSlip = [
-                        { url: bankSlipURL }
-                    ]
-                }
-                const financialSummaryId = await createFinancialSummarie(financialSummaryPayload);
-            } else {
-                console.log("Inside else")
-                console.log("Error occurred while uploading image");
+        }, (error, result, response) => {
+            if (error) {
                 console.log(error);
+                throw Error("Unable to upload blob from text");
             }
-        })
+        });
+
+        return bankSlipURL
+
     } catch (err) {
         console.log(err);
         throw Error("Unable to upload blob")
